hot-fix

+++ 0x01-variables_if_else_while/5-print_numbers.c
@@ -8,7 +8,7 @@
 
 int main(void)
 {
-double b = '0';
+int b = '0';
 
 /* finding value of log10 */
 while (/* condition */b < 10)
